Add default styling
